optionally dump effectively used tile, grid and block sizes
As explained in the README, it can be difficult to predict how
the statements will be distributed over the kernels, but also
what the dimension of the tilable band will be that is tiled
and mapped to blocks and threads.  The README recommends the user
to first run PPCG with the defaults and to then examine the kernels.
This can be difficult to automate, however.
A dump of the effectively used tile, grrid and block sizes
is easier to analyze.

Requested-by: "Betts, Adam" <a.betts@imperial.ac.uk>
Tested-by: "Betts, Adam" <a.betts@imperial.ac.uk>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>