gpu: insert kernel guard node in schedule tree
In the next commit, we will be introducing a context node in the schedule
tree that introduces the block and thread identifiers with their bounds.
The bounds on the block identifiers are parametric and may therefore
implicitly impose bounds on those parameters.
We insert a guard first to ensure that those implicit constraints
are enforced by the generated AST.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>