add --opencl-include-file option
In the next commit, we will pull in changes in pet that make it
possible to specify function summaries.  In order to be able
to exploit this functionality, we need to be able to tell the
OpenCL compiler where to find the function bodies for the functions
that are being called (and for which summary functions may be
specified).

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>