opencl backend: print kernel file using single printer
opencl_print_kernel() was creating and releasing an isl printer on each
invocation.  This makes postprocessing the kernel file contents a bit
difficult, which we need to do in a future commit.

Instead of writing directly to the kernel file, we now buffer the kernel
file contents using an isl string printer, and write it after all work
has been done.

Signed-off-by: Sven van Haastregt <sven.vanhaastregt@arm.com>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>