add --opencl-embed-kernel-code option
With this option, the kernel code is written to the kernel.cl file as a
string literal.  The kernel.cl file is then #included in the host file
such that the kernel.cl file is not needed at run time anymore.

The kernel string literal may become large if there are many kernels,
and some compilers may impose a limit on the size of string literals.
However, Clang (2.9, 3.4) and gcc (4.4, 4.6, 4.7) seem to compile a
128-kilobyte string literal without any problems.

Signed-off-by: Sven van Haastregt <sven.vanhaastregt@arm.com>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>