gpu: read tile sizes before ppcg_kernel objects are created
We want to perform the tiling that creates two bands with parallel dimensions
from a single band before the kernel creation such that in the future
we could also allow different ways of specifying two such bands.
This means that we need to read the tile sizes before the kernel objects
are created.

Since we currently still need those tile sizes during the AST generation,
we also save them in the ppcg_kernel object.  In order to get a handle
on the ppcg_kernel object, we temporarily store it in gen->kernel.
This will be removed as soon as we no longer need to store extra
information in the ppcg_kernel object inside mark_outer_tilable.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>