gpu: move force_private from gpu_array_info to gpu_local_array_info
The gpu_array_info elements are shared by the entire gpu_prog,
while the decision of whether to forcibly put arrays into registers
is kernel specific.
We currently use this flag before we consider the next kernel,
overwriting the old flag.
In the future, however, we will be constructing a single schedule tree
for the entire program, containing information about all kernels
together.  We therefore need to put this information in kernel
specific data structures.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>