opencl.c: move up opencl_copy_arrays_from_device
In some distant future we will be performing the copying as a special
type of user statement.  This function therefore needs to be placed
before opencl_print_host_user.
Since we will be modifying opencl_copy_arrays_from_device in the next
commit, we might as well move the function up now.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>