gpu: remove encoding of sequences in flat schedule representation
Now that we no longer use the host part of the flat schedule for
AST generation, there is no longer any need to perform these manipulations.

This commit may affect the computed grid and block sizes simply
because the order of the basic sets in the input of extract_size
may change.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>