ppcg_kernel: also keep constraints on outer schedule dimensions in context
In particular, postpone projecting out the outer schedule dimension
until the point where we need a context in terms of only the parameters,
i.e., when we are computing the effective grid size.
We will be able to reuse the context with outer schedule dimensions
when we introduce a kernel guard in the schedule tree in a later commit.
In the future, we may even want to take into account the outer schedule
dimensions during the computation of the effective grid size as well.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>