gpu: stop mapping array to shared memory if no consecutive elements are accessed
An access that never accesses consecutive elements cannot be coalesced,
at least not using the current scheme, so there is no point in mapping
the array to shared memory for the sole purpose of improving coalescing.
Check whether an access accesses consecutive memory elements and, if not,
consider the access as coalesced already so that the non-coalesced
property will not cause the array to be mapped to shared memory.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>