gpu: only preserve data that needs to be preserved
When determining what to copy into the device, we add those elements
that will be copied back from the device to the host, but that may
not have been written by the device in between.
If, however, some elements cannot possibly be preserved by the scop,
then there is no need to add them to the elements that need to be
copied in.  This happens in particular if these elements are
definitely killed inside the scop.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>