gpu.c: extract_access: extract access->access from access->tagged_access
access->access is just access->tagged_access with the tag removed,
so we might as well compute it like that instead of performing
essentially the same computation twice.
The first step is to extract access->access from access->tagged_access.
The duplicate computation will be removed in the next commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>