gpu.c: collect_array_info: collect all array references by scop
In particular, also collect arrays that may not be accessed at all.
This is a first step at supporting index expressions that refer
to an array but that do not end up accessing anything, e.g., because
they are passed to a function and the function doesn't use it.
We still need to be aware of the existence of such arrays.

The original code dates back from before we switched to pet.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>