gpu: handle disappearing localized array sizes
The array sizes are localized within each instance of a launched kernel
in order to simplify the size expressions.  Some of the arrays that
are accessed by the statements that are mapped to this kernel may
not actually be accessed by a given instance in case the AST generator
has performed separation resulting in a given kernel being launched
several times.  In this case, the simplified size expressions will
simply be undefined and we cannot print them.  Replace the sizes
by zero instead.  Since the access will never actually be performed
by this kernel instance, it does not matter which array sizes we use.

Reported-by: Michael Kruse <reply@meinersbur.de>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>