gpu: avoid copying data to/from global memory that is local to statement
Ever since c27d9a3 (avoid copying data to/from global memory that is
only used internally, Fri Feb 22 16:53:08 2013 +0100), we try and avoid
copying data to/from global memory that is known to be only used
inside a single instance of a kernel, by removing those data elements
that are only used in dataflow within such an instance.

However, in order not to needlessly break up the set of accessed
elements into smaller pieces, we only perform this restriction
if there are actually any data elements involved in local dataflow.
Unfortunately, this means the optimization is also skipped
if the accesses are not involved in any dataflow.
In particular, this may happen if the dataflow only occurs
within the same (compound) statement instance.

Try and handle those cases to some extent by at least removing
accesses from statements that are not involved in any dataflow.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>