ppcg.c: eliminate_dead_code: also intersect range of dataflow dependences
While removing local dataflow, we should also not consider the dataflow
to statement instances that have been eliminated as non-local dataflow.
Instead, it should simply not be considered as dataflow at all.
Remove the eliminated statement instances from the range of the dataflow
as soon as they have been eliminated.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>