use ppcg_scop_generate_names for block and thread identifiers
This ensures that the block and thread identifiers do not clash
with variable names in the scop.
The block and thread identifiers are now also constructed only
once and stored in ppcg_scop.
Some of the auxiliary functions are changed to take a list of isl_ids
instead of (re)generating names.  Since these functions are also
used by access_is_bijective, this function is also changed to use
ppcg_scop_generate_names.

With these changes, all variables created by PPCG are now made
not to clash with variable names in the scop.  We can therefore
remove the warning about possible clashes from the README.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>