opencl: replace calls to mathf functions by overloaded function and cast
OpenCL does not have any expf, powf or sqrtf functions but instead
provides overloaded exp, pow and sqrt functions.  If the *f functions
appear in the input code, then they need to be replaced by a cast to float
and a call to the corresponding overloaded function.

This is needed for the upcoming PolyBench/C 4.0.  We currently only handle
those functions used by PolyBench.  Other functions can be added later.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>