remove clang from blacklist for openmp support
In ce4c754 (polybench_test.sh.in: do not perform openmp tests when using
clang, Thu Feb 5 13:10:48 2015 +0100) clang was blacklisted because
there were some issues with some PolyBench benchmarks.  However, it turned
out that the problem was not caused by clang, but by a problem in
the initialization of the input for those benchmarks.  In particular,
part of the array was not being initialized and while the contents
apparently defaulted to zero on most systems, this failed to be
the case when adding the -fopenmp flag on the clang command line.
Refrain from blacklisting clang.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>