add dependences between sources with shared sinks to forced dependences
When live-range reordering is turned on, we ignore the usual false
dependences and replace them by order dependences between live-ranges
that are only activated when the adjacent live-ranges are not local
to a given partial schedule.
If there are any may-writes, however, then there may be several
live-ranges with the same sink (read), but different sources (writes).
The false dependences between these writes were also getting ignored,
while they should be enforced unconditionally since several of
the potential writes may turn out to actually write to the data element.

Compute dependences between such sources with shared sinks and
and them to the forced dependences such that they will be enforced
unconditionally.

Reported-by: Michael Kruse <MichaelKruse@meinersbur.de>
Tested-by: Michael Kruse <MichaelKruse@meinersbur.de>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>