exploit independences during dependence analysis
In e57e679 (gpu.c: compute_schedule: exploit independences,
Mon Feb 3 17:57:58 2014 +0100), we started exploiting independences,
but only very conservatively.  In particular, we would only use
them to allow more schedule dimensions to be considered parallel and
not to create more freedom for the scheduler.

Also exploit the independences during dependence analysis,
which does create more freedom for the scheduler.
We still only exploit independences when live-range reordering
is enabled since the live-range reordering takes care
of ensuring that the live-ranges of local variables will not overlap.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>