gpu backend: declare local variables that are used on the host
Variables that are local to the scop are currently not being declared
because the code is either entirely mapped to the device or
sent to the cpu backend.  Since there cannot be any dataflow into
or out of a local variable from outside of the scop, these variables
should not have to be copied in or out of the device.
In future, however, we will allow some part(s) of the scop
to be executed on the host.  If any local array ends up getting
copied, then we need to make sure it is declared on the host.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>