ppcg_kernel_requires_array_argument: check if kernel acceses global memory
That is, only require a kernel argument for a given array if that
kernel accesses the global memory corresponding to that array
rather than if the kernel accesses the array at all.
In particular, if the array is mapped to private or shared memory
without getting copied in or out, then no corresponding kernel
argument is needed.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>