ppcg.c: compute_forced_dependences: avoid use of before relation
The before relation contains all pairs of statement instances
where the first is executed before the second according to
the original schedule.  Its size is therefore quadratic in
the number of statements.

Remove the final use of the before relation as well as the computation
of the before relation itself.
Do so by computing the external false dependences directly from
the original schedule.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>