Allow original user statements to carry pointers
PPCG recognizes original user statements by the fact that their ids do
not carry a pointer value. To make PPCG more robust against future changes,
we want it to not depend on ids not carrying a pointer value. Instead of
assuming user statements are only carrying NULL pointers, we now check
in the list of statements if there is a gpu statement that corresponds
to the given isl id.

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>