gpu_group.c: access_is_coalesced: handle zero-dimensional data
The function access_is_coalesced is only called when there is no reuse.
If the array is zero-dimensional, then this means it is accessed
exactly once.  This is fairly unlikely, but it can happen if
PPCG gets confused about the amount of available parallelism.
Although in such cases, the produced code would be suboptimal,
we should at least produce some code instead of erroring out.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>