gpu_array_tile: keep track of own depth
The depth field of a gpu_array_ref_group is used for two purposes:
- the depth at which overlapping accesses should be considered
- the depth at which copying of a tile should be introduced

If both a shared and a private tile have been computed,
then to ensure correctness the minimal depth of the shared and
the private tile should be used to determine overlapping accesses.

Due to the double use of the group->depth field,
it is not clear if the current implementation will always get it right.
This commit therefore splits up the field in a group->min_depth
field used for determining overlap and a tile->depth field
to keep track of the copying depth for a particular tile.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>