gpu_array_ref_group_type: take tile with minimal depth if both types exist
If both a shared and a private tile can be computed, then the private tile
is preferred.  However, if the shared tile can be copied at a lower depth,
then the shared tile may be preferable.
This change is unlikely to make a difference at this point, but a future
commit will allow the mapping to shared memory to be computed at
a different initial depth.  If it is possible to compute such a shared
tile, then we do not want it to be discarded in favor of a private tile
that is computed deeper in the tree.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>