gpu_array_ref_group_compute_tiling: do not depend on dimension of group->access
The mapping from group->depth schedule dimensions to the array space
is constructed from the space of group->access, thereby assuming that
the input dimension of the latter is at least as large as group->depth.
Although this is currently the case and there are no immediate
plans for changing this, there is no need to make such an assumption.
Construct the space in a more obvious way.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>