only print macro definitions once
A macro definition for isl_ast_op_fdiv_q is currently printed unconditionally
because it may occur inside the AST expressions for array or grid sizes.
By telling the isl AST printer to only print macro definitions once,
the printing of the isl_ast_op_fdiv_q macro definition can be delayed
until the point where it is needed without ending up with multiple
definitions if it is needed several times.
This will also get rid of the special casing for isl_ast_op_fdiv_q,
which is currently the only operation that is expected in
such size expressions, but other operations may also appear
in those expressions in the future.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>