add ppcg_build_size_expr
In subsequent commits, some parametric size expressions will be converted
to AST expressions within the context of an isl_ast_build.
This requires the domain of the expression to match the schedule
space of the isl_ast_build.  The function ppcg_build_size_expr takes care of
first transforming the parametric size expression to an expression
with such a domain.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>