optionally use macro definitions for min/max with GNU extensions
The default isl macro definitions evaluate their arguments
several times.  This is not a problem in terms of correctness
because the arguments should not have any side-effects.
However, the compiler may not be able to figure out that
the expressions only need to be evaluated once.
By using GNU extensions, a macro definition can be written
that explicitly tells the compiler to only evaluate the arguments once.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>