gpu.c: extract_grid_size: coalesce grid instead of coalescing grid size
The grid is the result of a projection, so it may very well
have partly overlapping pieces.  Performing coalescing on this set
can result in fewer calls to isl_set_dim_max and will then typically
result in simpler output.  Coalescing this output should then no longer
be required.  Besides, it will be coalesced anyway when it is
converted to an AST expression.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>