gpu.c: accessed_by_domain: clarify nature of domain argument
In particular, clarify that the domain corresponds to
the original statement instances that appear in the domains
of the access relations.
The callers of accessed_by_domain pass in a set of statement
instances that are active at a given node in the schedule tree.
If this schedule tree contains any expansions, then these instances
may be different from those at the leaves.  Adjust the domain argument
of accessed_by_domain calls to take into account any possible expansion.
Note that the schedule trees from which the statement instances
are extracted currently contain at most expansions outside of
the node from which the instances are extracted, but this may
change when statement instances are combined during scheduling.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>