gpu.c: remove_local_accesses: clarify nature of "sched" argument
In particular, clarify that the domain of this schedule corresponds to
the original statement instances that appear in the domains
of the access relations.
One of the callers of remove_local_accesses (add_to_from_device) already
passes in a schedule of right form, but anchored_non_local_accesses
passes in a schedule of which the domain corresponds to the
instances that are active at a given node in the schedule tree.
If this schedule tree contains any expansions, then these instances
may be different from those at the leaves.  Adjust the schedule argument
of the remove_local_accesses call to take into account any possible expansion.
Note that the schedule trees from which the statement instances
are extracted currently contain at most expansions outside of
the node from which the instances are extracted, but this may
change when statement instances are combined during scheduling.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>