gpu_group.c: check_scalar_live_ranges: use gpu_array_can_map_to_private
check_scalar_live_ranges_in_host would use a slightly different test
from the one used in collect_order_dependences to determine whether
an array can be mapped to private memory.  It was still safe, because
it used a weaker test and would therefore at worst check some arrays
redundantly.  Use the same check to clarify the connection and
to make it easier to change the decision of which arrays can be mapped
to private memory.
Also rename the function to reflect the fact that any array that
may be mapped to private memory is considered.  While currently
only scalars may be mapped to private memory, this will be changed
in an upcoming commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>