gpu: allow any array accessed through constant index expressions to be private
The arrays that are considered for mapping to private memory directly
were restricted to scalars (0-dimensional arrays without subfields).
The reason is that no addressing is required to access them
and an array element that is accessed through a non-constant
index expression is automatically spilled to (slow) local memory in CUDA.
Extend the cases of arrays that are considered for mapping
to private memory to all those arrays that are only accessed
through accesses with a fixed, complete index expression.

This removes the need for adding the corresponding order dependences
to the coincidence constraints, increasing the scheduling freedom
in those cases where additional arrays are considered for mapping
to private memory.

Requested-by: Michael Kruse <MichaelKruse@meinersbur.de>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>